refactor de servicios

+++ grails-app/services/arduito/arduino/RegistroService.groovy
@@ -12,9 +12,9 @@ class RegistroService {
 
     def registrarMedicion(sensor,fecha,valorMedido) {
 
-		sensor.valorActual = valorMedido
+		sensor.valorActual = valorMedido as Float
 		sensor.save(flush:true)
-		def registro = new RegistroSensor(fechaCreacion:fecha,sensor:sensor,valorRegistrado:valorMedido)
+		def registro = new RegistroSensor(fechaCreacion:fecha,sensor:sensor,valorRegistrado:valorMedido as Float)
 		registro.save(flush:true)
     }
 }
